New utility: interface helper. (#735)
A utility is added in common/utilities: an interface helper. This is intended to allow interface types, basically jump tables, to be created and destroyed easily. u_interface.h is kept intentionally simple, relying only on stdint.h types, so that it can be included in other interface header files without a risk of dragging in still other headers. It is used/tested in the up-coming virtual serial interface type.