New device feature: virtual serial interface. (#736)
This commit introduces a new device type, a virtual serial interface. This may be used with cellular, short-range or GNSS devices and is a way to present an underlying serial-type interface (e.g. UART or USB) as a virtualised serial interface, following the pattern of u_port_uart.h. It is required to support the up-coming cellular 3GPP 27.010 CMUX feature, where a single UART to a cellular module turns into multiple serial interfaces using an HDLC-like multiplexing protocol; those interfaces can then be used either for AT commands or for transparent access to a GNSS chip embedded inside or connected via the cellular module.

This feature works with the aid of the recently introduced u_interface utility functions.