AT client improvements: what goes in can be read and hijack. (#747)
In the up-coming 3GPP 27.010 CMUX implementation it is necessary for an existing AT client to be copied onto a new transport (i.e. from the UART transport to a virtualised CMUX transport). Hence with this commit we add functions to read all settings that need to be copied.

This commit also adds the ability to "hijack" the URC handler; in the CMUX case, the UART interface that is underneath the AT client goes from carrying AT commands to carrying a weird HDLC-like multiplexing protocol, which makes absolutely no sense to an AT client, hence it must be possible to handle that outside the AT client.