HTTP fix for cellular only: port number selection.
If no port number was provided in the URL -1 (65535) would end up being used.

Now, if no port number is provided in the URL then no port number is provided to the cellular module (which will adopt the default of 80) and then, if security is switched on, the port is changed to 443 and, if security is switched off, the port number is changed to 80; if the caller _did_ specify a port number in their URL then the port number is left alone.