Fix working of uPortUartEventIsCallback(). (#745)
The function should check if the task that is currently running is the UART event callback thread. However, what it does is check that the UART event callback function has been called, not that it is currently running, so if the UART event callback function has yielded for whatever reason (e.g. it is waiting at a mutex lock or has called the RTOS wait function) the uPortUartEventIsCallback() function will still return that it is in the function, even though it is not. This PR fixes that.

Internal jira reference: https://jira.u-blox.net/browse/PGM_KM_AWS-397