Short range fix only: avoid EDM stream mutex lock-up when closing. (#771)
There is a potential mutex lock-up in the close down of a short-range EDM stream.  It goes like this:
- uShortRangeClose() calls uShortRangeEdmStreamClose(),
- uShortRangeEdmStreamClose() locks the uShorRange API gMutex,
- then it calls uPortEventQueueClose(),
- the task at the end of that event queue has to run to completion in order for the event queue to close,
- in this particular case, that task is generating EDM serial port events to the AT Client URC handler, which will call uShortRangeEdmStreamAtRead() to read the serial data out of EDM,
- uShortRangeEdmStreamAtRead() needs to lock the uShorRange API gMutex, which is already locked; we're stuck.

Since we already have a global flag gEdmStream.ignoreUartCallback, checked in uartCallback(), this fix checks the same flag in uShortRangeEdmStreamAtRead(), so that it returns immediately if the flag is true.  Using a global flag in this way is not the ideal answer but, since there can be only one EDM stream at the moment, it is sufficient to get us working; better than the lock-up.