UART fixes for STM32F4 and Windows: do not miss events. (#738)
The STM32F4 and Windows UART drivers included a flag, "userNeedsNotify", intended to stop needless events being issued. This flag was set to false when an event was queued and then set to true again when the UART read() or get-size() functions returned zero. However, use of a such a flag suffers from race conditions and it was possible for events to stop being sent. While the system would recover, the data is not lost, the AT client relies on such events to alert it to unsolicited result codes (e.g. +CEREG etc.) in a timely manner; if events aren't generated such URCs will only be processed when the AT client goes to send the next AT command, quite possibly after it has timed out waiting for a URC it really wanted (e.g. one containing an MQTT command completion URC).

With this commit the "userNeedsNotify" flag is removed; instead we suffer a slightly higher load, which we already do on all other platforms without issue, and STM32F4/Windows are our fastest platforms anyway.