Cellular fix, sockets, SARA-R422: what to do if no +USOWR. (#773)
We have seen cases where SARA-R422 responds to a AT+USOWR with just OK, not +USOWR: 12 or whatever to indicate the number of characters written.  At the moment the code will misbehave when it sees this, since the AT interpreter has seen "OK", so all is fine, yet the "number of bytes sent" parameter is missing, so comes back as -1.

What do to when this happens?  Has the data been sent or hasn't it, does "OK" mean it's all OK, or does the lack of a number of bytes sent mean no data was sent?  Now, in the cases I've seen this happen, it is the former, the data has been sent, but that still seems a risky assumption to make, so with this change we at least prevent the -1 happening and set the number of sent bytes to 0.

This might mean that data gets sent twice, but whaddaya do?