BREAKING MQTT change: callback ONLY IF NUMBER INCREASES. (#743)
Reading an MQTT message results in the module emitting a URC with the new number of unread messages in it, which will of course call the callback again.  Though irritating, this is normally of no consequence _except_ in the case where the module is returning from loss of coverage when a load of queued MQTT messages may suddenly arrive all at once; depending on the rate of arrival versus rate of reading a flurry of unnecessary callbacks/message-queuing may result, even leading to queue exhaustion in extreme cases.

With this commit the MQTT unread messages callback will only be called if the number of unread MQTT messages is greater than or equal to that of the last URC indication from the module and _not_ if the number of unread MQTT messages has gone down.