Add AT client URC priority in CMakeLists.txt (#766)
This change is required to support a fix for when the URC task queue fills up because a burst of URCs arrive at once, the fix being to raise one task's priority versus another, rather than keeping all of the queue/task priority linkages at the same priority.  Such a fix is tested and used on SARA-R5 AWS, where the application code behaves in a known way and hence such testing is possible.

An alternative fix, in which the task priorities can remain the same, is merged in PR #734.  Leaving the task priorities the same is preferred for the following reason:

queue 1 --> Task A --> queue 2 --> Task B --> queue 3 --> Task C

If the tasks run at the same priority the queues will balance each other out as the task scheduling is "even", no queue should ever fill up unless something is taking processing time, which is fair game, that's just life.

If Task B is made lower priority, queue 2 will fill up (because Task B is not emptying it fast enough) and Task A will be unable to complete because it will be stuck waiting for Task B to do its job. OK, you say, queue 2 has to be big enough to cope with the maximum number of URCs that will ever arrive at once: but what is that number?

Moreover, if Task B sends to another task, Task C, there is now a kink in the priority chain: Tasks A and C get to run, task B doesn't and you have a lock-up. Task C is the application, which of course runs at application priority, there's nothing that can done about that.

Putting it another way, RTOS priority-inversion detection mechanisms simply aren't good enough to get out of this situation, they need a little helping-hand from the code itself.