Test Automation change only: USB fix to RPi setup again. (#774)
The previous documentation change for the Raspberry Pi instances to switch a USB device off, and have it stay off, was insufficient: turns out that /sys/bus/usb/devices/1-1.x/authorized is volatile, it is reset when the USB device powers up again, so setting it to zero has to be done in the Jenkinsfile scripted pipeline.  Also the C030 and STM32F407 Discovery boards need the disk they map to be powered-down before the USB can be switched off.