Cellular and AT client change: service recovery improvement and uAtClientLockExtend(). (#739)
The cellular network reactivation process, performed asynchronously by ubxlib on return from a coverage outage, has to carry out a few AT exchanges with the module to do its work and locks/unlocks the AT client for each exchange. This is fine except that there may be other entities running at the same priority which also want to perform AT operations, and those may require service to have been reacquired before they will succeed, hence being scheduled by the OS part-way through (when the AT client is unlocked/relocked) is counter productive.

This commit introduces a new AT client API, uAtClientLockExtend(). This may be used in such situations, i.e. where you need to lock the AT client at the start of a sequence of operations and not give up that lock until the entire sequence has completed. You can replace the uAtClientUnlock()/uAtClientLock() pairs in such an AT client sequence with a call to uAtClientLockExtend() and it will reset the lock time and clear any errors just as uAtClientUnlock() would while leaving the AT client locked.

This function is used in the context activation code in order to ensure that it completes fully before others can get in.