Windows platform fix: UART "corruption". (#779)
It turns out that software flow control on a Windows serial port defaults to on, at least in some cases (FTDI drivers), even though Device Manager says that flow control is off for that serial port. And if software flow control is on, any 0x11 characters in the input stream will go missing. D'oh.

Software flow control is now forcibly switched off, with a vengeance, for Windows UARTs.