BREAKING port/include-path change: add SPI transport and common type API. (#751)
An SPI API is added in the porting layer and implemented on the Zephyr, ESP-IDF/[Arduino], NRF5SDK and STM32F4 platforms.  It is tested for use with u-blox GNSS chips.

IF YOU ARE NOT USING the provided ubxlib.cmake or ubxlib.make files, you will need to add a NEW INCLUDE PATH, common/type/api, to your build.  This include directory is for type definitions that straddle layers of ubxlib.  It currently contains common_spi.h, which is needed by both u_port_spi.h and u_device.h.

IF YOU HAVE YOUR OWN PORTING LAYER you can return U_ERROR_COMMON_NOT_SUPPORTED for all of the functions in the port SPI API without any adverse effect, unless of course you happen to want to use it to talk with a GNSS chip.

Note:  as part of this change the Zephyr portion of the SPI configuration, specifically the chip-select bit, now works with the Zephyr version 3 configuration structure used in NRF Connect 2.1.0.

Note about the runner build the Nordic MCUs: the runner build we employ for testing and running examples uses HW block SPI3 for nRF52 and HW block SPI2 for nRF53; using SPI2 for nRF53 (the natural choice since that's what the main Zephyr DTS file sets up) means that the UARTE HW block has to be shifted to be UARTE3.  None of this makes any change to the actual pins used though.