MEMORY CLEAN-UP CHANGE in the uPortEventQueue API: do not free memory immediately on close. (#776)
uPortEventQueueClose() used to wait for the task on the end of the event queue to complete all its actions before closing that event queue and freeing memory. This is problematic since uPortEventQueueClose() will lock its API mutex and the task on the end of the event queue might, potentially, call back into the uPortEventQueue API (e.g. the AT client URC handler does), leading to a lock-up on the API mutex.

Instead uPortEventQueueClose() now marks the event queue as closed and only frees memory if it runs out of room, or when event queues are deinitialised, or when the new uPortEventQueueCleanUp() function is called, all of which will happen long after the task has finished up its queue and gone quiet..

IMPORTANT: this means that if you call uDeviceClose(), uBleDeinit(), uCellDeinit(), uShortRangeEdmStreamClose(), uPortUartClose(), uAtClientDeinit(), uWifiMqttClose() or uMqttClientClose() not all heap memory will be returned to you (not until uPortDeinit() is called); if you wish to get that heap memory back immediately, call uPortEventQueueCleanUp().