Fix warning with logging off in GNSS and Wifi: unused variables. (#759)
In two places, one in GNSS and one in Wi-Fi, a variable was used only in the logging code and hence when logging is compiled out a warning was emitted. With this commit the "build with logging off" tests is updated to include a check for unused variables and the two occurrences are fixed.