TLS security fix for cellular: ignore port on SNI. (#761)
When a server name check (AKA SNI) is requested during TLS security it is possible that, for other reasons, the caller has included a port number on the end of the server string. The cellular module does not expect a port number on the end of the SNI string so uCellSecTlsSniSet() now removes any port that may be included.