Test workaround for nRF5SDK: halve data length in device interface test. (#787)
The test of the device interface type creates a UART interface that is just the usual UART physical interface abstracted through an interface type and loops data through that UART interface. For some reason, only on the nRF5SDK platform, only on test instance 15.0.0, this gets stuck writing to the UART and ends up mutex-locking itself.

There is a known issue with the UART interfaces of the ubxlib port layer, in that the uPortUartRead() and uPortUartWrite() interfaces both lock their API mutex; in this looped-back-to-self situation, there is always the potential for a mutex lock-up if the receive buffer is small enough that flow control switches off towards the "write" side before it is able to complete, with it keeping the mutex locked so that the "read" side cannot empty the buffer.

However, this device interface test is [deliberately] constructed in exactly the same way as the uPortUartRequiresSpecificWiring() test of the port tests, and the uPortUartRequiresSpecificWiring() test does not lock up at all, ever.

All a bit of a mystery really; however, since the interface test is not a test of the UARTs per se, this change reduces the transmit amount by half in the hope that it will, at least, not get stuck.