Documentation change, SARA-R422 security credentials: add workaround. (#792)
Issue https://github.com/u-blox/ubxlib/issues/77 points out that a SARA-R422-00B/01B module, when fresh from the factory, may fail to store a certificate.  This is a known bug, fixed in the SARA-R422-02B version, and can be worked-around by issuing uCellFileListFirst()/uCellFileListLast() before writing the very first certificate.

Many thanks to Johannesmay for raising and debugging this.