Generic BLE APIs added (#797)
* Generic BLE APIs added

This is a first implementation of more generic BLE APIs in
ubxlib. Both GAP and GATT level functionality is available.
The Nordic Uart Service (NUS) is also provided and built on
top of the new APIs.

The test sequence added here is for the NUS service as this
implicitly tests the GAP and GATT layers.
The test process needs two external BLE modules and firmware
for these is provided. This is for Nina B3 evks.

The old BLE SPS functionality is still available but can not
be used at the same time as the new APIs.