GNSS change only: active antenna configuration. (#804)
Two functions are added, uGnssCfgGetAntennaActive() and uGnssCfgSetAntennaActive(), to control whether the antenna is supplied with DC to power it or not.

The uGnssCfgSetAntennaActive() function is called in the preamble of each GNSS test so as to switch off DC bias to the antenna: this is because, on the test system, we connect all of the GNSS antenna outputs together to a separately-powered active antenna, hence we don't need the DC power from each device and leaving it on means they end up powering each other. However, it is possible to disable this automatic-disabling during testing with the flag U_GNSS_TEST_DISABLE_ACTIVE_ANTENNA_DISABLE; this is applied for the SARA-R422M8S cellular module of test instance 22 because, for some reason, it fails to gain position if the volts are taken away.