BREAKING CHANGE to uDeviceSerial_t: correct name for "DiscardOnOverflow" APIs. (#807)
Commit be7a1db16051cbaf28fc170841b1a4f8932b4dee introduced two new API types in uDeviceSerial_t: uDeviceSerialDiscardOnOverflow_t and uDeviceSerialIsDiscardOnOverflowEnabled_t.  However, in the implementation the naming got confused and, in the _function_ names, the word "Overflow" got replaced with "FlowControl", which is of course incorrect: discarding only happens when the buffer overflows, not when some arbitrary flow control limit happens to be reached.

This commit corrects the function names to discardOnOverflow and isDiscardOnOverflowEnabled, to match the types.  The version number of the interface (U_DEVICE_SERIAL_VERSION) is updated from 1 to 2, though no other attempt is made towards backwards compatibility; what went before was just plain wrong.

IF YOU ARE USING THE discardOnFlowControl and isDiscardOnFlowControlEnabled functions in the `uDeviceSerial_t` API, please update your code to use the discardOnOverflow and isDiscardOnOverflowEnabled functions instead.