Improvement in uSock: permit UDP uSockWrite()/uSockRead(). (#809)
There is no logical reason why uSockWrite()/uSockRead() calls should be prevented for UDP - while it is conventional that socket write/read functions are used for TCP and socket write-to/read-from functions are used for UDP, the only real difference, at this level, is that the socket has been connected and hence there is no need for the function call to include the destination address. Connection of a UDP socket is supported - no transaction occurs with the peer entity, there is no "TCP"ness, the local client just remembers the destination address.

With this commit the restriction that a call into uSockWrite()/uSockRead() must be for a TCP socket is removed.

Note that, for sockets over Wifi, calling uSockWrite()/uSockRead() on a UDP socket will still result in an error since uConnectExpress only supports use of uSockWrite()/uSockRead() on a TCP socket but, with this commit, it is possible to use uSockWrite()/uSockRead() over cellular on a UDP socket.