Cellular change only: uCellSockGetBytesPending(). (#814)
An API, uCellSockGetBytesPending() is added to get the number of bytes currently pending on a socket in the cellular module. This is useful if you wish to extend the time that the module spends in power saving mode since, unlike calling uCellSockRead(), it does not send an AT command to the module (which would wake the module up); instead it returns the number of bytes known to be available by this code based on the URCs previously emitted by the module.