Fix compiler warning in GNSS. (#816)
In uGnssPrivateStopMsgReceive() we send a message in order to shut-down a task, the contents of the message being irrelevant, just the arrival of the message is sufficient. Since we don't initialise the message contents to anything some compilers may warn that the variable is uninitialised, which of course it is. With this commit we initialise the message array to 0 to shut it up.