Update the test system to use nrfconnect 2.2.0 (#811)
* Update the test system to use nrfconnect 2.1.0

All tests running on Nordic based boards using Zephyr will now use
version 2.2.0 instead of the old 1.6.1. As this implicitly means using
Zephyr version > 3 all involved board device tree overlay files had to be
updated to a new syntax.
The nrf5340pdk board overlay has been removed as it is not used. The ninab3
one has been reintroduced as it is used in the Visual Studio build tasks
included in the workspace.
The git submodule of the u-blox board repo has been removed as it is no
longer used and as this repo has not been updated to Zephyr 3.

BLE is removed for NinaB4 for now as this needed to avoid flash size overflow

With nRFConnect 2/Zephyr 3 it is no longer possible to determine whether flow control pins are wired-up or not at compile time, hence the macros U_CFG_TEST_PIN_UART_A_xxx_GET and U_CFG_APP_PIN_UART_A_xxx_GET are now removed from the Zephyr version of u_cfg_test_platform_specific.h/u_cfg_app_platform_specific.h and the tests portUartRequiresSpecificWiring() and cellInfoImeiEtc() are modified to work without them.

In later versions of Zephyr there is a check if subscription has
been made when sending a BLE nofification event. This was not the
case before and the uport implementation doesn't set up this correctly.
This will make one of the test fail. As this check doesn't really matter
for now an option, which disables it is now set. A major overhaul of
opencpu BLE is due in coming versions.

Increased Jenkins timeout for flashing.

The test failures in the latest commit are not related to this PR.

---------

Co-authored-by: RobMeades <rob@meades.org>