Support is added for 3GPP 27.010 AT+CMUX, allowing multiple UARTs to be run across a single cellular UART interface; for instance an AT interface plus a GNSS UART interface that is "directly" connected to a GNSS module that is inside, or connected-via, the cellular module. (#808)
This mechanism is used by the uNetwork API in preference to using the AT+UGUBX command to exchange single UBX messages with a GNSS network [chip] inside/connected-via a cellular device; a subsequence commit will add a streamed position API to the GNSS/uLocation APIs to take advantage of the fact that all routes to GNSS can now be streamed.

Note that the heap required to use AT+CMUX with a GNSS channel is relatively large: 5 to 10 kbytes more than otherwise. This is because an additional AT client, buffers, etc. are required to get the AT and GNSS information from A to B. CMUX is ONLY invoked when a GNSS network is brought up on a cellular device; you may stick with the old AT+UGUBX ping/pong message mechanism by defining U_NETWORK_GNSS_CFG_CELL_USE_AT_ONLY.