Test automation change only: fix spelling in 53_dut.rules.
The 53_dut.rules file, used to map the USB devices (debuggers etc.) to named devices under Linux in our test automation, had a mistake in the "SEGGER JLink Base" line that mapped that debugger to a UART.  This didn't appear to have any detrimental effect _except_ that, when we introduced an nRF52833 board to the test system (instance 26), it would fail to flash-download intermittently (e.g. one in five times).  I can't, honestly, see why a mistake of this nature would cause an intermittent failure but, having spotted it, fixed it and copied it to the Raspberry Pi that is running instance 26, that instance succeeded in flashing the board 10 out of 10 times.

Of course, 'tis a statistical failure, so it will likely fail to flash download on the very next run after I've committed this, but the problem needed correcting anyway.  Fingers crossed.