Cellular fix for SARA-R422 only: corrected radio info for 2G fall-back. (#827)
The AT command AT+UCGED, used to retrieve information about cell ID/ARFCN, when used with a SARA-R422 cellular module, has different response formats for 2G versus Cat-M1/NB1. The Cat-M1/NB1 form was being assumed in all cases, meaning that if uCellInfoGetCellId()/uCellInfoGetEarfcn() were called when SARA-R422 happened to be in 2G-fallback, the wrong answer would be returned. This is now fixed.