MQTT improvement for both cell and Wifi: specific error on message read when empty.
To make it easier for an application to use the MQTT APIs, the functions uMqttClientMessageRead(), uCellMqttMessageRead() and uWifiMqttMessageRead() now return a specific error message, U_ERROR_COMMON_EMPTY, if a read is performed when there are no messages to be read.  As part of this change the internally common short-range function uShortRangePktListConsumePacket() will now return U_ERROR_COMMON_EMPTY if it is given an empty packet list.