Test change only for I2C port test: allow retries and disabling. (#828)
Testing of I2C is somewhat flaky: on some platforms (e.g. STM32F4 and Nordic) it simply seems to suffer from HW problems - either pull-ups or cable length or speed or interference or possibly the state that the GNSS chip at the other end of the I2C bus is in (though that chip is reset at the start of the test). Anyway, from now on we allow up to three goes to get the right answer back from the GNSS chip in the test portI2cRequiresSpecificWiring(), which will hopefully improve reliability.

...except that, even with 6 retries over a one minute period, our STM32F407 Discovery board still fails portI2cRequiresSpecificWiring(), getting zero bytes available reported by the GNSS chip when it should be 20. I have seen this work on a retry, but only once. Hence, for that particular platform (instance 19) the portI2cRequiresSpecificWiring() test is now disabled; to be clear, the same board/test instance still runs the true GNSS stuff over I2C perfectly happily, so I2C is still being tested, there's just something about the test scenario of portI2cRequiresSpecificWiring() it doesn't like.