BREAKING CHANGE for cellular: report RAT-appropriate SNR equivalent where available. (#835)
Up to now uCellInfoGetSnrDb() has reported SNR by determining the ratio of RSRP / (RSSI - RSRP) in dB for all RATs. However, that is really only appropriate for 2G; for 3G and 4G the appropriate SNR equivalent value, as reported directly by the module, is now returned. For 3G this is EC/N0, while for Cat-M1/NB/LTE it is the signal to interference and noise ratio (SINR).

EXISTING CELLULAR USERS may experience quite different [but now correct] SNR results for Cat-M1, NB, LTE and UTRAN.