GNSS streamed position fix only: retry putting settings back. (#837)
uGnssPosGetStreamedStart() sets the navigation measurement rate and the message rate of the UBX-NAV-PVT message when it is at work and uGnssPosGetStreamedStop() calls uGnssPrivateCleanUpStreamedPos() to put the original settings back again afterwards.  For some reason, when using a Virtual Serial port to an M8 chip embedded inside a cellular module (SARA-R510M8S or SARA-R422M8S), putting the settings back can fail; either the outgoing "set" command fails to get to the GNSS chip, or it succeeds in getting there (evidenced by reading the setting back later) but no ack comes back for it.  I've looked at the detailed CMUX debug in a failure scenario and there really is no sign of an ack message at all, nothing _appears_ to be being lost.  Hence, with this commit, putting the settings back as they were is retried up to two times on failure.