Cellular fix in multiplexer: remove double-discard. (#838)
If a multiplexer channel ended up having to discard a portion of an incoming I-field (e.g. as it sometimes has to for the GNSS channel on SARA-R5, which does not support CMUX flow control), it would discard it from the ring buffer and then _also_ discard the whole processed I-frame from the ring-buffer, so a double-discard, potentially losing wanted data from the ring buffer in the process. This is now fixed.