BREAKING CHANGE: fix for ESP-IDF platform linker workaround. (#844)
Commit bf8cd21ff293ba2e3cf2f750f682aff1cf335fc0 introduced a mechanism for leaving out one or more of cell, GNSS or short-range [BLE and Wifi] from a build.  As noted at the time, there is a bug in the ESP-IDF linker which means that, if a file happens to contain functions which ALL have WEAK counterparts, the linker will entirely ignore the .c file and use the WEAK versions instead; this is counter-productive, resulting in everything returning "NOT SUPPORTED".

A workaround for this problem was included in the commit: a dummy function with no WEAK counterpart was introduced to those files which, when called from somewhere, would cause the ESP-IDF linker to correctly link the non-WEAK functions.  However, the places we chose to call the _key_ dummy functions from (the functions for the uDevice and uNetwork APIs which provide the "root" of the linker tree) were, unfortunately, themselves functions that had WEAK counterparts, so the workaround kind of ate itself.

This is now fixed: those key dummy functions are now _always_ called and separate "_link.c" files are brought in, with the implementation of those dummy functions, for the cases where the code is left out (to keep all of the linkers happy).

IMPORTANT, THE BREAKING PART: if you bring the ubxlib files into your build in your own way (so not using ubxlib.mk or ubxlib.cmake or PlatformIO) and you are leaving out one or more of cellular, GNSS or short-range [BLE and Wifi], you MUST NOW BRING INTO YOUR BUILD the "_link.c" files for the part that you are _leaving_out_, otherwise linking will fail.

Note: the non-PlatformIO-flavour-Arduino-build does not support leaving bits out; please use the PlatformIO Arduino build if you need that capability.