Cellular MQTT fix: publish in low signal conditions with keep-alive. (#846)
When publishing an MQTT message, if MQTT keep-alive is on and a ping message has been sent to the broker by the module just before the publish request commences, the ">" prompt for a binary send will not be returned by the module until the ping completes or times out. In fringe conditions it can take up to 30 seconds for the ping to succeed or timeout. Hence, if keep-alive is on, the timeout for the MQTT publish is extended by U_CELL_MQTT_PROMPT_TIMEOUT_KEEP_ALIVE_SECONDS (30) in addition to the U_MQTT_CLIENT_RESPONSE_WAIT_SECONDS (30) now allowed for the publish itself.