Test automation change for Zephyr: build as a Zephyr module. (#847)
The runner and runner_linux builds for Zephyr include Zephyr using the "ZEPHYR_EXTRA_MODULES" mechanism but, of course, they bring in their own specific configuration for testing.  This builds Zephyr as a module using the minimum viable prj.conf, with an empty application, to be sure it compiles/links.