BREAKING CHANGE for ESP32 S3: I2C timeout not supported. (#851)
On the S3 version of the ESP32 chip the I2C timeout register has gone from 20 bits in length, so max value 0xFFFFF, which at an 80 MHz APB clock is about 13 milliseconds, to 5 bits in length, so max value 31 APB clock cycles which is just 0.38 nanoseconds. No idea why this is (I have asked) but it means that we can't practically support setting an N millisecond timeout for the S3 case.

With this commit, uPortI2cSetTimeout() and uPortI2cGetTimeout() will return U_ERROR_COMMON_NOT_SUPPORTED for the ESP32 S3 chip.