Test change only: fix compilation warning in test httpClient(). (#853)
In GCC 7.1 and later, when a length-limited string function (e.g. snprintf()) is used and the compiler spots that (a) there may be a truncation and (b) the caller does not check the return value of the function in order to do something about it, it generates a warning. This occurs in the test httpClient() where we know there could be truncation, that's why we're using snprintf(), and we don't care.

Tried checking the return value and then (void)'ing it to show we don't care but that didn't fix it, so have instead limited the length of the string in the printf() formatter.