ESP-IDF fix for cellular HTTP: one more linker workaround (#860)
Commit bf8cd21ff293ba2e3cf2f750f682aff1cf335fc0 introduced a mechanism for leaving out one or more of cell, GNSS or short-range [BLE and Wifi] from a build.  This hit bug in the ESP-IDF linker which means that, if a file happens to contain functions which ALL have WEAK counterparts, the linker will entirely ignore the .c file and use the WEAK versions instead; this is counter-productive, resulting in everything returning "NOT SUPPORTED".

This linker bug was worked-around in commit 6eff2c4b669d2b42f85a7488304323012272c589 but it turns out that even that missed a bit, in the cellular HTTP code, where the file system calls needed a dummy "link" function to get them included.  Hopefully this is the last of them.