Change to inactivity debug printing, plus tidy-ups.
With the up-coming changes in timeouts for cellular, the potential inactivity times in _normal_ operation will increase, hence the inactivity time for the debug utility which, on ESP32 and Zephyr, can be enabled to print out task stacks (by defining U_DEBUG_UTILS_DUMP_THREADS), is hereby increased from 60 seconds to 180 seconds.

Also included are a few tidy-ups in cellular and the MQTT client.