Cellular fix, CMUX: switch echo off on CMUX port.
With some modules (e.g. SARA-R5) whether echo is on or off can be port-specific and hence echo needs to be switched off on the CMUX port when it is created.

Note: this is a weird one.  CMUX has been working without issue on SARA-R5 and then, last week, one of the SARA-R5 modules on the test system [running 03.15 IP FW as usual] began exhibiting issues because echo would be ON on the CMUX port, even though it was OFF on the non-CMUX ort.  Before then having echo off on the non-CMUX port would also mean echo off on the CMUX port.  There have been no code changes in ubxlib or on the cellular module, so I can't explain why this is happening now, but issuing ATE0 on the CMUX port definitely fixes it.