Test change only: update deprecated OpenOCD commands for STM32F4. (#873)
Some of the OpenOCD commands used by the STM32F4 part of the test system (hal_serial and tpiu) are deprecated; the relevant Python script is now updated to use the correct commands.  Also the adapter type for the STLink V2 standalone debugger, which we might use in future, is added to the udev rules list.