Added WiFi captive portal function (#872)
* Added WiFi captive portal function

The WiFi captive portal is mainly intended to provide a simple
way to do provisioning of WiFi credentials to a module in ubxlib.
In order to achieved this implementation of the previously missing
socket apis for server functionality had to be implemented. This
is done by using the server functions available in uConnect.
Apis for starting a WiFi module as an access point have also
been added in this commit.
Dns server functionality has been added as well.
New api for setting the hostname of a WiFi module.
WiFi credentials storage apis added.
Captive portal test function which relies on external client added.