BREAKING CHANGE for SPI: uCommonSpiControllerDevice_t gets a new member, indexSelect. (#875)
When we moved to Zephyr 3/NCS2 it became necessary for the SPI chip select pin to be specified both in the Zephyr device tree and to the ubxlib code, which is a bit strange for Zephyr users and is unnecessary duplication. With this change ubxlib only needs to know which of the chip-select pins defined in the Zephyr device tree is to be used as SPI chip select, it does NOT need to know the actual value, which is a bit easier to deal with.

The uCommonSpiControllerDevice_t has a new element, indexSelect: this points to the nth cs-gpios entry that you will have specified for the SPI controller in the Zephyr device tree. The U_COMMON_SPI_CONTROLLER_DEVICE_DEFAULTS() initialiser macro is updated to include this new field (setting indexSelect to -1, so no chip-select index by default, just like pinSelect is -1 by default) and a new initialiser macro, U_COMMON_SPI_CONTROLLER_DEVICE_INDEX_DEFAULTS() is added to accommodate use of indexSelect rather than a pin.

If pinSelect is negative and indexSelect is non-negative then the index will be used. If pinSelect and indexSelect are both negative then no chip-select will be employed.

Provided you have always assigned every SPI device structure to U_COMMON_SPI_CONTROLLER_DEVICE_DEFAULTS() at initialisation, you should notice no change.

Zephyr users: if your SPI controller includes one or more cs-gpios you may find it more convenient to use indexSelect/U_COMMON_SPI_CONTROLLER_DEVICE_INDEX_DEFAULTS() rather than pinSelect/U_COMMON_SPI_CONTROLLER_DEVICE_DEFAULTS(), though both will work.