ESP-IDF version update: support version 5. (#876)
The ESP-IDF version supported/tested by ubxlib was previously v4.3.2, though v4.4 was known to work (Espressif introduced support for ESP32S3 in v4.4, as used in NORA-W1).

With this commit support for ESP-IDF version 5 (5.0.1 to be exact) is added and the ubxlib test system has been moved to test with this version.  v4.3 (and v4.4 in fact) should continue to work, tested as part of the Arduino platform, which still uses ESP-IDF v4.3.

Note: a few warnings about an ignored "noreturn" attribute are being emitted from the depths of the ESP-IDF header files when compiling the file port.c.  After a discussion in https://github.com/espressif/esp-idf/issues/11339 this seems to be because, earlier in the includes, we bring in stdnoreturn.h, which _shouldn't_ be a problem but, right now, is.  This PR does nothing to suppress the warnings: we will step forward an ESP-IDF version when Espressif have made a fix, or we will address it with a change to the conditionals in compiler.h if they decide that what they have is perfectly fine.