Examples change only: don't uDeviceClose() unless uDeviceOpen()ed. (#887)
The examples all assume that uDeviceOpen() will succeed, not checking the return code, just to keep things simple and clear. However, quite often when a user is first figuring things out uDeviceOpen() may fail (due to wrong pin settings etc. in the configuration) and, when that happens, the examples can go "bang" trying to close an uninitialised pointer to a device structure, which looks a bit pants.

Hence with this commit the examples are all updated to only close a device that has actually opened.